+++ CHANGELOG.md
@@ -3,6 +3,10 @@ This is the change log for the plugin, all relevant changes will be listed here.
 
 For documentation please see the [README](https://github.com/normen/homebridge-bravia/blob/master/README.md)
 
+## 2.4.6
+- Store channels for external TVs in file as the device cache in homebridge doesn't store external devices
+- Improve channel update logic
+
 ## 2.4.5
 - Allow enabling debug output per TV through UI
 - Fix debug output not working